fixed issues with test functions

+++ SimISR/analysisplots.py
@@ -322,7 +322,7 @@ def plotbeamparametersv2(times,configfile,maindir,fitdir = 'Fitted',params=['Ne'
     for i_fig in range(nfig):
         lines = [None]*2
         labels = [None]*2
-        (figmplf, axmat) = plt.subplots(sp.ceil(Np/2), 2,figsize=(20, 15), facecolor='w')
+        (figmplf, axmat) = plt.subplots(int(sp.ceil(Np/2)), 2,figsize=(20, 15), facecolor='w')
         axvec = axmat.flatten()
         # loop that goes through each axis loops through each parameter, beam
         # then time.
